Move istio specific functions in seprate files
