+++ pkg/appgw/configbuilder.go
@@ -12,7 +12,6 @@ import (
 	n "github.com/Azure/azure-sdk-for-go/services/network/mgmt/2018-12-01/network"
 	"github.com/Azure/go-autorest/autorest/to"
 	"github.com/golang/glog"
-	"github.com/knative/pkg/apis/istio/v1alpha3"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/api/extensions/v1beta1"
 	"k8s.io/client-go/tools/record"
@@ -148,27 +147,6 @@ func (c *appGwConfigBuilder) resolvePortName(portName string, backendID *backend
 	return resolvedPorts
 }
 
-func (c *appGwConfigBuilder) resolveIstioPortName(portName string, destinationID *istioDestinationIdentifier) map[int32]interface{} {
-	resolvedPorts := make(map[int32]interface{})
-	endpoints, err := c.k8sContext.GetEndpointsByService(destinationID.serviceKey())
-	if err != nil {
-		glog.Error("Could not fetch endpoint by service key from cache", err)
-		return resolvedPorts
-	}
-
-	if endpoints == nil {
-		return resolvedPorts
-	}
-	for _, subset := range endpoints.Subsets {
-		for _, epPort := range subset.Ports {
-			if epPort.Name == portName {
-				resolvedPorts[epPort.Port] = nil
-			}
-		}
-	}
-	return resolvedPorts
-}
-
 func generateBackendID(ingress *v1beta1.Ingress, rule *v1beta1.IngressRule, path *v1beta1.HTTPIngressPath, backend *v1beta1.IngressBackend) backendIdentifier {
 	return backendIdentifier{
 		serviceIdentifier: serviceIdentifier{
@@ -179,28 +157,6 @@ func generateBackendID(ingress *v1beta1.Ingress, rule *v1beta1.IngressRule, path
 		Rule:    rule,
 		Path:    path,
 		Backend: backend,
-	}
-}
-
-func generateIstioMatchID(virtualService *v1alpha3.VirtualService, rule *v1alpha3.HTTPRoute, match *v1alpha3.HTTPMatchRequest, destinations []*v1alpha3.Destination) istioMatchIdentifier {
-	return istioMatchIdentifier{
-		Namespace:      virtualService.Namespace,
-		VirtualService: virtualService,
-		Rule:           rule,
-		Match:          match,
-		Destinations:   destinations,
-		Gateways:       match.Gateways,
-	}
-}
-
-func generateIstioDestinationID(virtualService *v1alpha3.VirtualService, destination *v1alpha3.Destination) istioDestinationIdentifier {
-	return istioDestinationIdentifier{
-		serviceIdentifier: serviceIdentifier{
-			Namespace: virtualService.Namespace,
-			Name:      destination.Host,
-		},
-		VirtualService: virtualService,
-		Destination:    destination,
 	}
 }
